Login y otros

+++ backend/FBProject/Models/Producto.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string? nombre_p { get; set; }
         public string? codigo_p { get; set; }
-        public decimal valor_p { get; set; }
+        public long valor_p { get; set; }
         public int unidades_p { get; set; }
         public DateTime created_at { get; set; }
